Move handling of announcement exception to app exception handler class
RISDEV-4710
